Version 1.2 final changes

+++ backend/src/main/java/com/fribbels/core/StatCalculator.java
@@ -15,10 +15,10 @@ public class StatCalculator {
         int hp = (int) (((base.getHp()   + mapAccumulatorArrsToFloat(1, accs)  + (sets[0] > 1 ? sets[0] / 2 * 0.15f * base.getHp() : 0) + base.getHp() * hero.getBonusHpPercent() / 100f) + hero.getBonusHp()) * (1 + base.getBonusMaxHpPercent()/100f));
         int def = (int) (((base.getDef() + mapAccumulatorArrsToFloat(2, accs)  + (sets[1] > 1 ? sets[1] / 2 * 0.15f * base.getDef() : 0) + base.getDef() * hero.getBonusDefPercent() / 100f) + hero.getBonusDef()) * (1 + base.getBonusMaxDefPercent()/100f));
         int spd = (int) (base.getSpd() + mapAccumulatorArrsToFloat(10, accs) + (sets[3] > 1 ? sets[3] / 4 * 0.25f * base.getSpd() : 0) + (sets[14] > 1 ? sets[14] / 4 * 0.1f * base.getSpd() : 0)) + hero.getBonusSpeed();
-        int cr = (int) (base.getCr()   + mapAccumulatorArrsToFloat(6, accs)  + (sets[4] > 1 ? sets[4] / 2 * 12 : 0)) + hero.getBonusCr();
-        int cd = (int) (base.getCd()   + mapAccumulatorArrsToFloat(7, accs)  + (sets[6] > 1 ? sets[6] / 4 * 40 : 0)) + hero.getBonusCd();
-        int eff = (int) (base.getEff() + mapAccumulatorArrsToFloat(8, accs)  + (sets[5] > 1 ? sets[5] / 2 * 20 : 0)) + hero.getBonusEff();
-        int res = (int) (base.getRes() + mapAccumulatorArrsToFloat(9, accs)  + (sets[9] > 1 ? sets[9] / 2 * 20 : 0)) + hero.getBonusRes();
+        int cr = (int) (base.getCr()   + mapAccumulatorArrsToFloat(6, accs)  + (sets[4] > 1 ? sets[4] / 2 * 12 : 0)) + (int) hero.getBonusCr();
+        int cd = (int) (base.getCd()   + mapAccumulatorArrsToFloat(7, accs)  + (sets[6] > 1 ? sets[6] / 4 * 40 : 0)) + (int) hero.getBonusCd();
+        int eff = (int) (base.getEff() + mapAccumulatorArrsToFloat(8, accs)  + (sets[5] > 1 ? sets[5] / 2 * 20 : 0)) + (int) hero.getBonusEff();
+        int res = (int) (base.getRes() + mapAccumulatorArrsToFloat(9, accs)  + (sets[9] > 1 ? sets[9] / 2 * 20 : 0)) + (int) hero.getBonusRes();
         int dac = base.getDac() + sets[10] / 2 * 4;
 
         float critRate = Math.min(cr, 100) / 100f;
