+++ backend/src/main/java/com/fribbels/request/BonusStatsRequest.java
@@ -15,14 +15,14 @@ public class BonusStatsRequest extends Request {
     private int atk;
     private int def;
     private int hp;
-    private int atkPercent;
-    private int defPercent;
-    private int hpPercent;
+    private float atkPercent;
+    private float defPercent;
+    private float hpPercent;
     private int speed;
-    private int cr;
-    private int cd;
-    private int eff;
-    private int res;
+    private float cr;
+    private float cd;
+    private float eff;
+    private float res;
     private String heroId;
 }
